refactor(database): Create module for database administration
