+++ src/app.module.ts
@@ -1,7 +1,6 @@
 import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
 import { APP_FILTER, APP_GUARD } from '@nestjs/core';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { TypeOrmModule } from "@nestjs/typeorm";
+import { ConfigModule } from '@nestjs/config';
 
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
@@ -9,8 +8,7 @@ import { CatsModule } from './cats/cats.module';
 import { HttpExceptionFilter } from './common/filters/httpException.filter';
 import { RolesGuard } from './common/guards/roles.guard';
 import { validate } from "./config/env.validation";
-import { Cat } from './cats/cat.entity';
-import { ENVIRONMENT } from './config/constants/env.constants';
+import { DatabaseModule } from './core/database/database.module';
 
 
 @Module({
@@ -20,20 +18,7 @@ import { ENVIRONMENT } from './config/constants/env.constants';
       isGlobal: true,
       validate
     }),
-    TypeOrmModule.forRootAsync({
-      imports: [ConfigModule],
-      inject: [ConfigService],
-      useFactory: (configService: ConfigService) => ({
-        type: 'mysql',
-        host: configService.get(ENVIRONMENT.DATABASE.HOST),
-        port: +configService.get(ENVIRONMENT.DATABASE.PORT),
-        username: configService.get(ENVIRONMENT.DATABASE.USERNAME),
-        password: configService.get(ENVIRONMENT.DATABASE.PASSWORD),
-        database: configService.get(ENVIRONMENT.DATABASE.NAME),
-        entities: [Cat],
-        synchronize: false,
-      })
-    }),
+    DatabaseModule,
   ],
   controllers: [AppController],
   providers: [
