+++ src/cats/dto/ListAllEntities.dto.ts
