+++ src/core/database/entities/cat.entity.ts
